libva-utils: initial commit
libva-utils is a new git repository that will host sample applications
for libva, conformance tests and others relevant to libva

It is using the same version as libva and it should follow libva version.
It is intended to be released together with libva
It has the same license as libva

Initial functionality ports the vainfo application from libva repository.
The git history for it will remain on libva.  To port vainfo, also the
a common libva-display library was ported. The original tests for encoder
and decoder are also ported to this repository. debian folder for vainfo
was also ported

The basic tests are going to be substituted by a new set of tests based
on Gtest framework.

libva-utils will dynamically link libva, use pkg-config to compile and
link properly.

Signed-off-by: Daniel Charles <daniel.charles@intel.com>