gtest: initial Gtest set for VA-API
This initial set of files covers the va.h API for all the
functions used to exercise successfully the VA-API.

Tests are divided per VA-API function with a common Fixture that
contains all calling functions to the VA-API.  For the use cases
covered already there are tests for the proper functionality and also
for expected error situations.

The tests creates combinations with all the inputs to the VA-API function
and then it checks the results based on what the driver implementation reports
as supported

The set of tests can be run on any h/w supported and any driver.

The VendorString test is assuming Intel i965 driver, please help
adding the test for other drivers

More tests will be continously added

Signed-off-by: Daniel Charles <daniel.charles@intel.com>