test: allow user-defined driver name for test execution
Some tests manipulate the LIBVA_DRIVER_NAME environment
variable for various reasons.  This causes tests that
execute later on to behave unexpectedly.

Store the original user-defined LIBVA_DRIVER_NAME before
every test runs and restore it to the original value
when each test finishes.  This allows the test suite
to be executed with any VAAPI driver.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>