tests: Return NULL if you can't get VADisplay
This patch should not change the outcome as m_vaDisplay is NULL for
each failed vaGetDisplayDRM().

This is minor code readability improvement. Another approach would be
to break the loop when we find a valid VADisplay.

While at it, remove the empty line before variable declaration which
is not common in the majority of the source code.

Signed-off-by: Victor Toso <victortoso@redhat.com>