vp9enc: Add SVCT Support
This patch adds supports 2 or 3 temporal layers, which can be activated
by using the '--temp_svc X' command line.
Note that the alternate reference frame is not used - this design
decission was carried on from libyami - in order to allow single
Layer 2 frames to be dropped independently.

This work is part of a GSOC 2018 project, mentored by Intel Media And
Audio For Linux.