Update to googletest 1.11.0.
This is commit e2239ee6043f73722e7aa812a459f54a28552929 from git.
This brings in SetUpTestSuite support.  Note that it required a few
minor updates to the tests mostly to quiet deprecation warnings, and
also to include a header we need that used to be brought in by gtest.